moved example files to example folder
