+++ example/configure_output_file_paths.py
